dns-client/requirements.txt
dnspython>=2.4.2
requests>=2.31.0
PyYAML>=6.0.1
cryptography>=41.0.7
pystray>=0.19.5
Pillow>=10.0.0